fix: resolve type errors in product position provider
- Fix int/double type mismatch in fallback layout calculation
- Change maxItemsInRow from double to int to match floor() return type
- Resolve CI/CD compilation errors

+++ lib/features/fridge/presentation/providers/product_position_provider.dart
@@ -243,7 +243,7 @@ List<_PlannedPlacement> _layoutSection(_SectionLayoutContext context) {
   );
 
   final double minHorizontalGap = 2.0;
-  final double maxItemsInRow = (availableWidth / (fallbackSize.width + minHorizontalGap)).floor();
+  final int maxItemsInRow = (availableWidth / (fallbackSize.width + minHorizontalGap)).floor();
   final int itemsToShow = math.min(products.length, maxItemsInRow);
 
   if (itemsToShow == 0) {
